refactor(completion): extract lexical context helpers into submodule (#0000)
